feat(dapp): schema file
